Split common and agent-specific dependencies (#182)
* Split LangGraph agent-specific dependencies

The added agents/langgraph/pyproject.toml defines a subproject for the
LangGraph-based CurrencyAgent and contains its dependencies. This commit
partially addresses issue #181.

Refs: issue #181

* Split sample google_adk agent and hosts deps

Adds pyproject.toml definitions to decouple dependencies in subprojects:
* agents/google_adk
* hosts/cli
* hosts/multiagent

Refs: issue #181, PR #182

* Remove agent-specific deps from common definitions

With this change adding samples/python to a sample agent or external
project will only add packages necessary for modules in "common" and
"utils". Together with the previous two commits,
* 0b85b58098030fb1582558f05c54897165c64785
* 33c258dcfe7b4931d6d0d8371bcaf7a8affba8a9
This closes issue #181.

Refs: issue #181, PR #182.

* Remove unused agents/langgraph/uv.lock

This is a leftover file from the development stage. In the latest config, only
uv.lock from the parent dir `samples/python`, is being used.

* Update run instructions in hosts/cli/README.md